chore: enable upcoming feature `InferSendableFromCaptures`
A massive number of concurrency warnings concern `KeyPath` and
issues with its sendability. Rather than chase these down make
use of SE-0418, "Inferring Sendable for Methods and KeyPath
Literals".

`swift-tools-version` is updated to 5.10, making the upcoming
feature, `InferSendableFromCaptures` available. Once done, the
compiler needs a little help assembling a correct `KeyPath` for
`\EnvironmentValues.theme`. A hint that it should be a
`WritableKeyPath` is enough to get it on track.